Temporary Fix for liquids.lua
Temporary fix for Liquids. There is a bug that causes liquids to be unable to be used as a crafting ingredient when they are at a negative temperature. This bug will be fixed in the next experimental release (0.14.4), but will not be fixed in the current 0.13.x version. These changes can be undone when/if the mod is updated to 0.14.x versions.